lib: Add support for tracking boot consoles
Add tty_list and boot_tty to the config struct to keep track of
available console interfaces and the default console to set as primary
respectively.
Also add a tty field to the boot_command struct so that the current
console can be sent to the discover server during a manual boot command.

Signed-off-by: Samuel Mendoza-Jonas <sam@mendozajonas.com>