discover: Add 'sync' user event
Add a user event named 'sync' that causes the discover server to merge
the devicemapper snapshots of mounted devices. This is particularly
useful as a debug aid (for example, when copying logs to a USB device),
as the server will otherwise only sync changes to mounted devices in
response to parser actions.

The command can be called as

	pb-event sync@device

to sync a particular device, or as

	pb-event sync@all

to sync all devices with snapshots.

Signed-off-by: Samuel Mendoza-Jonas <sam@mendozajonas.com>