ui/ncurses: Make server connect message more clear
The current message mentions a "server" which can give the misleading
impression that the UI is waiting for a remote network server. The delay
is actually in waiting for the pb-discover process to be ready, so
update the message to reflect that.

Signed-off-by: Samuel Mendoza-Jonas <sam@mendozajonas.com>
(cherry picked from commit ea8bd5f424a9b01c71e1694bb5b6ffc13e3da4f5)