discover/boot: Fix talloc parent for resource URLs
The pb_urls for the boot resources use 'opt' as the talloc parent but
this may be NULL, for example if this is a user-created boot option.
Move the boot_task initialisation so it can be used as the talloc parent
instead.

Signed-off-by: Samuel Mendoza-Jonas <sam@mendozajonas.com>
(cherry picked from commit ef0c91e9b6c99ccbb5a3c1a88bb4d5ca85d254b4)