discover/grub2: Allow using title for default even if id was defined
A default menu entry can be chosen using any of the following attributes:
index, title or id (if the entry was defined with the --id option).

But Petitboot doesn't honor this correctly and only compares the default
with the menu entry title if the entry doesn't have an id defined. This
is wrong since an index or title can be used even if an id was defined.

This issue wasn't covered by the test that sets a default using a title
because the menu entries didn't have an id defined. Add an id to them.

Signed-off-by: Javier Martinez Canillas <javierm@redhat.com>
Signed-off-by: Samuel Mendoza-Jonas <sam@mendozajonas.com>