pb-console: start shell as a login shell
The environment that pb-console starts is minimal, as we have likely
been run by udev.

Because we're starting a shell, we want the proper profile set up, so
run as a login shell.

This assumes the shell accepts -l, but we already have an assumption for
supporting -m (as /bin/ash does).

Signed-off-by: Jeremy Kerr <jk@ozlabs.org>